+++ test/compute-engine/awesome.test.ts
@@ -125,17 +125,9 @@ describe('Nth PRIME NUMBER', () =>
                                                 "Subtract",
                                                 [
                                                   "Floor",
-                                                  [
-                                                    "Divide",
-                                                    ["Subscript", "v", 2],
-                                                    ["Subscript", "v", 3]
-                                                  ]
+                                                  ["Divide", "v_2", "v_3"]
                                                 ],
-                                                [
-                                                  "Divide",
-                                                  ["Subscript", "v", 2],
-                                                  ["Subscript", "v", 3]
-                                                ]
+                                                ["Divide", "v_2", "v_3"]
                                               ]
                                             ]
                                           ]
@@ -143,22 +135,14 @@ describe('Nth PRIME NUMBER', () =>
                                       ],
                                       [
                                         "Triple",
-                                        "Nothing",
+                                        "v_3",
                                         2,
-                                        [
-                                          "Floor",
-                                          ["Sqrt", ["Subscript", "v", 2]]
-                                        ]
+                                        ["Floor", ["Sqrt", "v_2"]]
                                       ]
                                     ]
                                   ]
                                 ],
-                                [
-                                  "Triple",
-                                  "Nothing",
-                                  2,
-                                  ["Subscript", "v", 1]
-                                ]
+                                ["Triple", "v_2", 2, "v_1"]
                               ]
                             ]
                           ]
@@ -170,7 +154,7 @@ describe('Nth PRIME NUMBER', () =>
                 ],
                 [
                   "Triple",
-                  "Nothing",
+                  "v_1",
                   2,
                   ["Floor", ["Multiply", 1.5, "n", ["Ln", "n"]]]
                 ]
@@ -223,14 +207,9 @@ describe('Nth PRIME NUMBER', () =>
                                 ]
                               ]
                             ],
-                            [
-                              "Limits",
-                              "Nothing",
-                              2,
-                              ["Floor", ["Sqrt", "v_2"]]
-                            ]
+                            ["Limits", "v_3", 2, ["Floor", ["Sqrt", "v_2"]]]
                           ],
-                          ["Limits", "Nothing", 2, "v_1"]
+                          ["Limits", "v_2", 2, "v_1"]
                         ]
                       ]
                     ],
@@ -240,7 +219,7 @@ describe('Nth PRIME NUMBER', () =>
               ],
               [
                 "Limits",
-                "Nothing",
+                "v_1",
                 2,
                 ["Floor", ["Multiply", 1.5, "n", ["Ln", "n"]]]
               ]
@@ -250,7 +229,7 @@ describe('Nth PRIME NUMBER', () =>
           "n"
         ]
       ]
-      eval-auto = (n) |-> sum_(2)^(floor(1.5 * n * ln(n)))(floor(1 / (0^(n - sum_(2)^("v_1")(prod_(2)^(floor(sqrt("v_2")))(1 - 0^(|-"v_2" / "v_3" + floor("v_2" / "v_3")|)))) + 1))) + 2
+      eval-auto = (n) |-> sum_(v_1=2)^(floor(1.5 * n * ln(n)))(floor(1 / (0^(n - sum_(v_2=2)^("v_1")(prod_(v_3=2)^(floor(sqrt("v_2")))(1 - 0^(|-"v_2" / "v_3" + floor("v_2" / "v_3")|)))) + 1))) + 2
     `);
   }));